pyproject.toml
[project]
name = "mongo-cxx-driver"
version = "0.1.0"
description = "For development only."
requires-python = ">=3.10"
dependencies = []

[dependency-groups]
dev = [
    # {include-group="apidocs"},
    {include-group="format"},
    {include-group="config_generator"},
    # {include-group="make_release"},
]

apidocs = [
    # etc/patch-apidocs-current-redirects.py
    "packaging>=14.0",

    # etc/patch-apidocs-index-pages.py
    "beautifulsoup4>=4.12",
    "packaging>=14.0",
]

format = [
    "clang-format~=20.1.0",
    "ruff>=0.12.10",
]

config_generator = [
    # .evergreen/config_generator/generate.py
    "packaging>=14.0",
    "pydantic>=2.8",
    "shrub-py>=3.7",
]

make_release = [
    # etc/make_release.py
    "click>=6.0",
    "cmake>=4.0",
    "gitpython>=3.1",
    "jira>=3.1",
    "looseversion>=1.3",
    "pygithub>=2.1",
]

[tool.ruff]
line-length = 120
src = [".evergreen", "etc"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
